made sure shader is not reset in mats when game version is changed
