+++ addon/i3dio/__init__.py
@@ -45,6 +45,7 @@ def register():
     ui.helper_functions.register()
     ui.addon_preferences.register()
     ui.udim_picker.register()
+    ui.shader_parser.register()
     ui.shader_picker.register()
     ui.exporter.register()
     ui.bit_mask_editor.register()
@@ -66,6 +67,7 @@ def unregister():
     ui.mesh.unregister()
     ui.light.unregister()
     ui.shader_picker.unregister()
+    ui.shader_parser.unregister()
     ui.udim_picker.unregister()
     ui.addon_preferences.unregister()
     ui.helper_functions.unregister()
