+++ addon/i3dio/ui/shader_picker.py
@@ -1,5 +1,4 @@
 from __future__ import annotations
-from dataclasses import dataclass, field
 from pathlib import Path
 
 import bpy
@@ -13,55 +12,13 @@ from bpy.props import (
 )
 from bpy.app.handlers import (persistent, load_post)
 
-from .helper_functions import get_fs_data_path, detect_fs_version, is_version_compatible, humanize_template
-from .. import xml_i3d
+from .shader_parser import (get_shader_dict, ShaderMetadata, ShaderParameter, ShaderTexture)
+from .helper_functions import (get_fs_data_path, detect_fs_version, is_version_compatible, humanize_template)
 
 
-# A module value to represent what the field shows when a shader is not selected
 SHADER_NO_VARIATION = 'None'
 SHADER_PARAMETER_MAX_DECIMALS = 3  # 0-6 per blender properties documentation
-SHADER_DEFAULT = 'NO_SHADER'
-
-SHADERS_GAME: ShaderDict = {}
-SHADERS_CUSTOM: ShaderDict = {}
-SHADER_ENUM_ITEMS_DEFAULT = (f'{SHADER_DEFAULT}', 'No Shader', 'No Shader Selected')
-SHADER_ENUMS_GAME = [SHADER_ENUM_ITEMS_DEFAULT]
-SHADER_ENUMS_CUSTOM = [SHADER_ENUM_ITEMS_DEFAULT]
-
-
-@dataclass
-class ShaderParameter:
-    name: str
-    type: int
-    default_value: list[float]
-    min_value: float = -xml_i3d.i3d_max
-    max_value: float = xml_i3d.i3d_max
-    description: str = ''
-    template: str = 'default'
-
-
-@dataclass
-class ShaderTexture:
-    name: str
-    default_file: str
-    template: str = 'default'
-
-
-@dataclass
-class ShaderMetadata:
-    path: Path
-    variations: dict[str, list[str]] = field(default_factory=dict)
-    parameters: dict[str, list[ShaderParameter]] = field(default_factory=dict)
-    textures: dict[str, list[ShaderTexture]] = field(default_factory=dict)
-    vertex_attributes: dict[str, str] = field(default_factory=dict)
-    param_lookup: dict[str, ShaderParameter] = field(default_factory=dict)
-
-
-ShaderDict = dict[str, ShaderMetadata]
-
-
-def get_shader_dict(use_custom: bool) -> ShaderDict:
-    return SHADERS_CUSTOM if use_custom else SHADERS_GAME
+SHADER_DEFAULT = ''
 
 
 def _clone_shader_texture(tex: I3DShaderTexture) -> dict:
@@ -186,39 +143,8 @@ class I3DShaderVariation(bpy.types.PropertyGroup):
 
 @register
 class I3DMaterialShader(bpy.types.PropertyGroup):
-    def shader_items_update(self, _context) -> list[tuple[str, str, str]]:
-        return SHADER_ENUMS_CUSTOM if self.use_custom_shaders else SHADER_ENUMS_GAME
-
-    def shader_setter(self, selected_index: int) -> None:
-        existing_shader = self.get('shader', 0)
-        if existing_shader != selected_index:
-            self['shader'] = selected_index
-            shader_enums = SHADER_ENUMS_CUSTOM if self.use_custom_shaders else SHADER_ENUMS_GAME
-            shader_name = shader_enums[selected_index][0]
-            ShaderManager(self.id_data).update_shader(shader_name)
-
-    def shader_getter(self):
-        return self.get('shader', 0)
-
-    def noop_update(self, _context) -> None:
-        return
-
-    shader: EnumProperty(
-        name='Shader',
-        description='The shader to use for this material',
-        default=0,
-        items=shader_items_update,
-        update=noop_update,  # Required to ensure depsgraph sync after file load (if material is migrated)
-        get=shader_getter,
-        set=shader_setter,
-    )
-
-    # Just for easy access to the shader name
-    shader_name: StringProperty(name=SHADER_DEFAULT)
-
     def custom_shaders_update(self, _context) -> None:
-        self['shader'] = 0
-        self.shader_name = SHADER_DEFAULT
+        self['shader_name'] = SHADER_DEFAULT
         ShaderManager(self.id_data).update_shader(SHADER_DEFAULT)
 
     use_custom_shaders: BoolProperty(
@@ -226,6 +152,27 @@ class I3DMaterialShader(bpy.types.PropertyGroup):
         description='Use a custom shaders instead of the game shaders',
         default=False,
         update=custom_shaders_update
+    )
+
+    def _shader_name_getter(self) -> str:
+        return self.get('shader_name', "")
+
+    def _shader_name_setter(self, shader_name: str) -> None:
+        existing_shader = self.get('shader_name', SHADER_DEFAULT)
+        if existing_shader != shader_name:
+            self['shader_name'] = shader_name
+            ShaderManager(self.id_data).update_shader(shader_name)
+
+    def _shader_name_search(self, _context, _search: str) -> list[str]:
+        shader_dict = get_shader_dict(self.use_custom_shaders)
+        return [name for name in shader_dict]
+
+    shader_name: StringProperty(
+        name='Shader',
+        description='The shader to use for this material',
+        get=_shader_name_getter,
+        set=_shader_name_setter,
+        search=_shader_name_search
     )
 
     # Variations
@@ -248,7 +195,7 @@ class I3DMaterialShader(bpy.types.PropertyGroup):
         return self.get('shader_variation_name', SHADER_NO_VARIATION)
 
     shader_variation_name: StringProperty(
-        name="Selected Variation",
+        name="Variation",
         description="The selected variation for the current shader",
         default=SHADER_NO_VARIATION,
         get=variation_getter,
@@ -318,10 +265,13 @@ class I3D_IO_PT_material_shader(Panel):
             box.label(text="If you want to convert this shader to new format, run the operator")
 
         custom_shaders_col = col.column(align=True)
-        custom_shaders_col.enabled = len(context.scene.i3dio.shader_folders) > 0
+        custom_shaders_col.enabled = len(context.scene.i3dio.custom_shaders)
         custom_shaders_col.prop(i3d_attributes, 'use_custom_shaders')
-        col.prop(i3d_attributes, 'shader', text="Shader")
-        col.prop_search(i3d_attributes, 'shader_variation_name', i3d_attributes, 'shader_variations', text="Variation")
+
+        is_shader_default = i3d_attributes.shader_name == SHADER_DEFAULT
+        shader_icon = 'FILE_BLANK' if is_shader_default else 'FILE'
+        col.prop(i3d_attributes, 'shader_name', placeholder="No Shader", icon=shader_icon)
+        col.prop_search(i3d_attributes, 'shader_variation_name', i3d_attributes, 'shader_variations')
 
         if i3d_attributes.required_vertex_attributes:
             column = layout.column(align=True)
@@ -332,7 +282,7 @@ class I3D_IO_PT_material_shader(Panel):
                 row.label(text=attr.name, icon='DOT')
             column.separator(factor=2.5, type='LINE')
 
-        if i3d_attributes.shader_name not in (SHADER_DEFAULT, ""):
+        if not is_shader_default:
             draw_shader_group_panels(layout, i3d_attributes)
 
 
@@ -389,146 +339,6 @@ def draw_shader_group_panels(layout: bpy.types.UILayout, i3d_attributes) -> None
         draw_shader_group_panel(layout, idname, group_label, i3d_attributes, params, textures)
 
 
-def parse_shader_parameters(parameter: xml_i3d.XML_Element) -> list[ShaderParameter]:
-    """Parses a shader parameter element and returns a list of dictionaries with parameter data."""
-    parameter_list: list[ShaderParameter] = []
-
-    type_str = parameter.attrib.get('type', 'float4')
-    type_length = {'float': 1, 'float1': 1, 'float2': 2, 'float3': 3, 'float4': 4}.get(type_str, 4)
-
-    def _parse_floats(val: str | None, default: float = 0.0) -> list[float]:
-        if val is None:
-            return [default] * type_length
-        try:
-            vals = [float(x) for x in val.split()]
-            # If too many, truncate; if too few, pad with default
-            return vals[:type_length] + [default] * (type_length - len(vals))
-        except Exception:
-            return [default] * type_length
-
-    param_name = parameter.attrib['name']
-    template = parameter.attrib.get('template', 'default')
-    default_value = _parse_floats(parameter.attrib.get('defaultValue'))
-    min_str = parameter.attrib.get('minValue')
-    max_str = parameter.attrib.get('maxValue')
-    min_value = _parse_floats(min_str) if min_str else [min(-xml_i3d.i3d_max, min(default_value))] * type_length
-    max_value = _parse_floats(max_str) if max_str else [max(xml_i3d.i3d_max, max(default_value))] * type_length
-    # Blender supports only a single min/max per prop, so if all are the same, use that; else fallback to i3d_max
-    min_single = min_value[0] if all(x == min_value[0] for x in min_value) else -xml_i3d.i3d_max
-    max_single = max_value[0] if all(x == max_value[0] for x in max_value) else xml_i3d.i3d_max
-
-    description = parameter.attrib.get('description', '')
-    if parameter.attrib.get('arraySize') is not None:
-        for child in parameter:
-            child_default = _parse_floats(child.text)
-            parameter_list.append(ShaderParameter(
-                name=f"{param_name}{child.attrib.get('index', '')}",
-                type=type_length,
-                default_value=child_default,
-                min_value=min_single,
-                max_value=max_single,
-                description=description,
-                template=template
-            ))
-    else:
-        parameter_list.append(ShaderParameter(
-            name=param_name,
-            type=type_length,
-            default_value=default_value,
-            min_value=min_single,
-            max_value=max_single,
-            description=description,
-            template=template
-        ))
-
-    return parameter_list
-
-
-def parse_shader_texture(texture: xml_i3d.XML_Element) -> ShaderTexture:
-    """Parses a shader texture element and returns a dictionary with texture data."""
-    return ShaderTexture(name=texture.attrib['name'], default_file=texture.attrib.get('defaultFilename', ''),
-                         template=texture.attrib.get('template', 'default'))
-
-
-def load_shader(path: Path) -> ShaderMetadata | None:
-    tree = xml_i3d.parse(path)
-    if tree is None:
-        return None
-    root = tree.getroot()
-    if root.tag != 'CustomShader':
-        return None
-    shader = ShaderMetadata(path)
-
-    if (variations := root.find('Variations')) is not None:
-        for v in variations:
-            if v.tag == 'Variation':
-                # Some variations don't have a group defined, but should still use the 'base' group regardless
-                shader.variations[v.attrib.get('name')] = v.attrib.get('groups', 'base').split()
-
-    if (parameters := root.find('Parameters')) is not None:
-        for p in parameters:
-            if p.tag == 'Parameter':  # Default to "base" if no group is specified
-                shader.parameters.setdefault(p.attrib.get('group', 'base'), []).extend(parse_shader_parameters(p))
-
-    if (textures := root.find('Textures')) is not None:
-        for t in textures:
-            if t.tag == 'Texture':  # Default to "base" if no group is specified
-                shader.textures.setdefault(t.attrib.get('group', 'base'), []).append(parse_shader_texture(t))
-
-    if (vertex_attributes := root.find('VertexAttributes')) is not None:
-        for attr in vertex_attributes:
-            if attr.tag == 'VertexAttribute':
-                shader.vertex_attributes[attr.attrib['name']] = attr.attrib.get('group', 'base')
-
-    # Add a lookup for parameters to easily access them by name
-    shader.param_lookup = {param.name: param for group in shader.parameters.values() for param in group}
-    return shader
-
-
-def load_shaders_from_directory(directory: Path) -> dict:
-    """Scans a directory for .xml shader files and returns a dict of shader_name -> ShaderMetadata"""
-    return {path.stem: shader for path in directory.glob('*.xml') if (shader := load_shader(path))}
-
-
-def populate_game_shaders() -> None:
-    global SHADERS_GAME, SHADER_ENUMS_GAME
-    SHADERS_GAME.clear()
-
-    shader_dir = get_fs_data_path(as_path=True) / 'shaders'
-    if shader_dir.exists():
-        SHADERS_GAME.update(load_shaders_from_directory(shader_dir))
-
-    SHADER_ENUMS_GAME = [SHADER_ENUM_ITEMS_DEFAULT]
-    SHADER_ENUMS_GAME.extend([(name, name, str(shader.path)) for name, shader in SHADERS_GAME.items()])
-    print(f"Loaded {len(SHADERS_GAME)} game shaders")
-
-
-def populate_custom_shaders() -> None:
-    global SHADERS_CUSTOM, SHADER_ENUMS_CUSTOM
-    SHADERS_CUSTOM.clear()
-
-    try:
-        for scene in bpy.data.scenes:
-            for entry in scene.i3dio.shader_folders:
-                path = Path(bpy.path.abspath(entry.path))
-                if path.exists():
-                    SHADERS_CUSTOM.update(load_shaders_from_directory(path))
-                else:
-                    print(f"[Custom Shader] Folder does not exist: {entry.path}")
-    except Exception as e:
-        print("Error reading custom shader folders:", e)
-
-    SHADER_ENUMS_CUSTOM = [SHADER_ENUM_ITEMS_DEFAULT]
-    SHADER_ENUMS_CUSTOM.extend([(name, name, str(shader.path)) for name, shader in SHADERS_CUSTOM.items()])
-    print(f"Loaded {len(SHADERS_CUSTOM)} custom shaders")
-
-
-@persistent
-def populate_shader_cache_handler(_dummy) -> None:
-    populate_game_shaders()
-    populate_custom_shaders()
-
-
 def _debug_print(message: str) -> None:
     if True:
         print(message)
@@ -540,10 +350,10 @@ def _migrate_shader_source(i3d_attr, old_shader_path: Path) -> bool:
     current_version = detect_fs_version(get_fs_data_path(as_path=True))
 
     # Check if the shader path matches any of the game shaders
-    if any(old_shader_path == s.path for s in SHADERS_GAME.values()):
+    if any(old_shader_path == s.path for s in get_shader_dict().values()):
         _debug_print(f"[ShaderUpgrade] Found game shader: {old_shader_stem} through path match")
         i3d_attr.shader = old_shader_stem
-    elif old_shader_stem in SHADERS_GAME:  # Check if the shader name matches any of the game shaders
+    elif old_shader_stem in get_shader_dict():  # Check if the shader name matches any of the game shaders
         if not is_version_compatible(old_version, current_version) and old_shader_stem == "vehicleShader":
             # Conversion for vehicleShader from 19/22 to 25 is a more involved process and need to be handled separately
             _debug_print(f"[ShaderUpgrade] Found game shader: {old_shader_stem} through name match, but not compatible")
@@ -552,8 +362,8 @@ def _migrate_shader_source(i3d_attr, old_shader_path: Path) -> bool:
         i3d_attr.shader = old_shader_stem
     elif old_shader_path.exists():  # We have to assume this is a custom shader
         _debug_print(f"[ShaderUpgrade] Found custom shader: {old_shader_stem} through path match to directory")
-        if old_shader_stem not in SHADERS_CUSTOM:
-            shader_folder_item = bpy.context.scene.i3dio.shader_folders.add()
+        if old_shader_stem not in get_shader_dict(True):
+            shader_folder_item = bpy.context.scene.i3dio.custom_shader_folders.add()
             shader_folder_item.name = old_shader_path.parent.stem
             shader_folder_item.path = str(old_shader_path.parent)  # Add folder where the shader is located
         i3d_attr.use_custom_shaders = True
@@ -648,13 +458,11 @@ def register():
     for cls in classes:
         bpy.utils.register_class(cls)
     bpy.types.Material.i3d_attributes = PointerProperty(type=I3DMaterialShader)
-    load_post.append(populate_shader_cache_handler)
     load_post.append(migrate_old_shader_format)
 
 
 def unregister():
     load_post.remove(migrate_old_shader_format)
-    load_post.remove(populate_shader_cache_handler)
     for cls in reversed(classes):
         bpy.utils.unregister_class(cls)
     del bpy.types.Material.i3d_attributes
